Update: new explain analyze content

+++ src/common/adbc/nanoarrow/allocator.cpp
@@ -23,11 +23,11 @@
 namespace duckdb_nanoarrow {
 
 void *ArrowMalloc(int64_t size) {
-	return malloc(size_t(size));
+	return malloc(size);
 }
 
 void *ArrowRealloc(void *ptr, int64_t size) {
-	return realloc(ptr, size_t(size));
+	return realloc(ptr, size);
 }
 
 void ArrowFree(void *ptr) {
